Move and rename internal interop classes to Interop folder
