+++ src/EventLogExpert.Eventing.Tests/EventResolvers/EventResolverBaseTests.cs
@@ -3,6 +3,7 @@
 
 using EventLogExpert.Eventing.EventResolvers;
 using EventLogExpert.Eventing.Helpers;
+using EventLogExpert.Eventing.Interop;
 using EventLogExpert.Eventing.Models;
 using EventLogExpert.Eventing.Providers;
 using EventLogExpert.Eventing.Tests.TestUtils;
