+++ src/EventLogExpert.Eventing.Tests/Helpers/NativeErrorResolverTests.cs
@@ -1,21 +1,21 @@
 // // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
+using EventLogExpert.Eventing.Interop;
 
 namespace EventLogExpert.Eventing.Tests.Helpers;
 
-public sealed class ResolverMethodsTests
+public sealed class NativeErrorResolverTests
 {
     [Fact]
     public void GetErrorMessage_ShouldReturnConsistentResults()
     {
         // Arrange — use a well-known HRESULT (ERROR_SUCCESS = 0)
-        const uint errorSuccess = 0;
+        const uint ErrorSuccess = 0;
 
         // Act
-        var result1 = ResolverMethods.GetErrorMessage(errorSuccess);
-        var result2 = ResolverMethods.GetErrorMessage(errorSuccess);
+        var result1 = NativeErrorResolver.GetErrorMessage(ErrorSuccess);
+        var result2 = NativeErrorResolver.GetErrorMessage(ErrorSuccess);
 
         // Assert — same string returned (cached)
         Assert.NotNull(result1);
@@ -26,16 +26,16 @@ public sealed class ResolverMethodsTests
     public void GetErrorMessage_WhenUnknownCode_ShouldReturnHexFallback()
     {
         // Arrange — use a code unlikely to have a system message
-        const uint unknownCode = 0xDEADBEEF;
+        const uint UnknownCode = 0xDEADBEEF;
 
         // Act
-        var result = ResolverMethods.GetErrorMessage(unknownCode);
+        var result = NativeErrorResolver.GetErrorMessage(UnknownCode);
 
         // Assert — if the OS doesn't resolve this code, we get the hex fallback.
         // On some OS versions/locales FormatMessage may resolve it, so only assert
         // the hex representation when the system has no message for this code.
-        var systemMessage = NativeMethods.FormatSystemMessage(unknownCode);
-        var ntStatusMessage = NativeMethods.FormatNtStatusMessage(unknownCode);
+        var systemMessage = NativeMethods.FormatSystemMessage(UnknownCode);
+        var ntStatusMessage = NativeMethods.FormatNtStatusMessage(UnknownCode);
 
         if (systemMessage is null && ntStatusMessage is null)
         {
@@ -52,11 +52,11 @@ public sealed class ResolverMethodsTests
     public void GetNtStatusMessage_ShouldReturnConsistentResults()
     {
         // Arrange — STATUS_SUCCESS = 0
-        const uint statusSuccess = 0;
+        const uint StatusSuccess = 0;
 
         // Act
-        var result1 = ResolverMethods.GetNtStatusMessage(statusSuccess);
-        var result2 = ResolverMethods.GetNtStatusMessage(statusSuccess);
+        var result1 = NativeErrorResolver.GetNtStatusMessage(StatusSuccess);
+        var result2 = NativeErrorResolver.GetNtStatusMessage(StatusSuccess);
 
         // Assert
         Assert.NotNull(result1);
@@ -67,6 +67,6 @@ public sealed class ResolverMethodsTests
     public void MaxCacheSize_ShouldBeReasonableBound()
     {
         // Assert — the cache limit is the expected constant
-        Assert.Equal(4096, ResolverMethods.MaxCacheSize);
+        Assert.Equal(4096, NativeErrorResolver.MaxCacheSize);
     }
 }
