+++ src/EventLogExpert.Eventing.Tests/Helpers/NativeMethodsTests.cs
@@ -1,7 +1,7 @@
 // // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
+using EventLogExpert.Eventing.Interop;
 
 namespace EventLogExpert.Eventing.Tests.Helpers;
