+++ src/EventLogExpert.Eventing/Interop/LibraryHandle.cs
@@ -1,7 +1,6 @@
 ﻿// // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
 using Microsoft.Win32.SafeHandles;
 
 namespace EventLogExpert.Eventing.Interop;
